refactor: Create BAse Inference model to simplify creation ONNX, OpenVino modules
