+++ configs/eval.yaml
@@ -3,7 +3,8 @@ mode: eval
 device: cuda
 pretrained: ./experiments/run_0/models/best_Model.pth
 metrics: {types: [accuracy_char, accuracy_word]}
-model: {backbone: ResNet ,sequence: BiLSTM,
+model: {backbone: SimpleNet ,sequence: BiLSTM,
+        pool: {name: AvgPool, factor: 6},
         prediction: CTC, input_channel: 3, output_channel: 512,
         hidden_size: 256, num_class: 2, num_fiducial: 20}
 data: {all_in_memory: True, datapath: ./data,
